requirements.txt
flask
flask-sqlalchemy
flask-cors
spacy
nltk
scikit-learn
numpy
pandas
werkzeug
python-docx
PyPDF2
faker
